refactor: 收口预览导出功能域
