+++ xl_updata_tool/app/features/preview/adapter.py
@@ -1,0 +1,35 @@
+"""Spine/FFmpeg 外部工具适配的 Preview Feature 入口。"""
+
+from app.ui.adapters.spine_adapter import (
+    composite_images,
+    composite_with_offset,
+    cleanup_temp,
+    export_animation_frames,
+    export_skel_skins,
+    export_spine_media_file,
+    extract_skin_name_from_png,
+    extract_character_id,
+    extract_motion_names,
+    ffmpeg_composite_videos,
+    find_composite_sources,
+    find_paired_files,
+    get_animation_names,
+    is_composite_png,
+)
+
+__all__ = [
+    "cleanup_temp",
+    "composite_images",
+    "composite_with_offset",
+    "export_animation_frames",
+    "export_skel_skins",
+    "export_spine_media_file",
+    "extract_character_id",
+    "extract_motion_names",
+    "extract_skin_name_from_png",
+    "ffmpeg_composite_videos",
+    "find_composite_sources",
+    "find_paired_files",
+    "get_animation_names",
+    "is_composite_png",
+]
