+++ xl_updata_tool/app/features/preview/fgui.py
@@ -1,0 +1,5 @@
+"""FGUI 图集切割的 Preview Feature 入口。"""
+
+from app.ui.features.fgui_atlas import UIPackageTool
+
+__all__ = ["UIPackageTool"]
