+++ xl_updata_tool/docs/代码所有权与边界.md
@@ -30,6 +30,7 @@
 11. Characters P2 完成后，角色页面信号、角色索引、Lua 解析协调、未读状态和 CSV 入口统一由 `app/features/characters/` 管理；MainWindow 只参与导航和导入结果接入。
 12. Versions P3 完成后，版本表格、版本选择、更新检查、下载计划、删除和磁盘 Bundle 状态统一由 `app/features/versions/` 管理；MainWindow 只参与导航、全局进度和导入选择适配。
 13. Importer P4 完成后，导出分类规则、精准 Bundle 筛选、导入结果和后处理分类统一由 `app/features/importer/` 管理；MainWindow 只启动控制器、展示全局进度并消费结果，`ImportASWorker` 仅作为兼容执行器保留。
+14. Preview P5 完成后，图片预览、FGUI、Spine/FFmpeg 导出编排和批量任务入口统一由 `app/features/preview/` 管理；`app/ui/features/export_controller.py`、旧 Worker 和旧适配器仅作为兼容底层，不在 UI Shell 中继续扩展领域流程。
 
 ## 提交边界
