+++ xl_updata_tool/docs/开发文档指南.md
@@ -283,6 +283,14 @@ SQLite 数据库管理，存储版本信息和资源包路径。
 - `app/features/preview/worker.py`：包装迁移期图片加载、预览导出、批量导出和合成导出 Worker，保持原信号/输出兼容。
 - FGUI 图集、Spine 适配器和视频批量导出仍处于 P5b 兼容迁移阶段，算法和旧调用入口暂不改变。
 
+#### Preview P5b：FGUI / Spine / 视频导出
+
+- `app/features/preview/export_controller.py`：拥有单文件、批量、合成图视频导出编排；旧 `app/ui/features/export_controller.py` 只做兼容转发。
+- `app/features/preview/fgui.py`：提供 FGUI `UIPackageTool` 的功能域入口；图集切割算法和 `output/fgui/` 契约保持不变。
+- `app/features/preview/adapter.py`：提供 Spine/FFmpeg 适配函数的功能域入口；外部工具调用参数、超时和输出路径保持原实现。
+- `app/features/preview/worker.py`：统一图片加载、预览导出、普通 Spine 批量导出和合成图视频导出 Worker 的功能域入口。
+- Platform/Shared 阶段再处理外部进程适配器的最终归属；本阶段不重写 Spine、FFmpeg 或 FGUI 算法。
+
 ### ui/ 子目录职责说明
 
 | 子目录 | 职责 | 主要文件 |
