+++ xl_updata_tool/docs/版本历史.md
@@ -32,6 +32,15 @@
 - 保留 `output/character/`、Spine 导出算法、旧页面/Worker 和视频/FGUI 导出入口；下一步继续 P5b 边界收口。
 - 新增 Preview 服务与架构边界测试，MainWindow 不再直接构造图片加载和预览导出 Worker。
 
+### v1.59.0（开发中）
+
+#### 架构迁移 P5b：FGUI / Spine / 视频导出入口（Issue #44）
+
+- 导出编排迁入 `app/features/preview/export_controller.py`，旧 UI 模块改为兼容转发。
+- 新增 Preview Feature 的 FGUI 入口、Spine/FFmpeg 适配入口和批量/合成 Worker 入口。
+- 保持角色图片、FGUI 图集、视频输出目录、导出参数和取消语义；外部工具底层算法暂不改写。
+- 下一阶段将继续收口 Platform/Shared 外部工具适配，避免 UI Feature 直接依赖旧基础设施路径。
+
 ### v1.55.0（开发中）
 
 #### 架构迁移 P2：Characters Feature ownership（Issue #41）
