+++ 开发计划与沟通事项.md
@@ -450,6 +450,17 @@ app/ui/
 - MainWindow 已切换图片预览页面、目录扫描、角色筛选、缩略图加载和立绘导出任务生命周期；旧 `preview_view.py`、旧 Worker 和导出控制器保留兼容。
 - 已完成定向测试、全量 pytest、Ruff、实际 Qt 启动冒烟与 diff 检查；用户验证通过后继续收口 FGUI、Spine 合成和视频导出。
 
+## Issue #44：Preview P5b FGUI/Spine/视频导出边界开始（2026-08-24）
+
+- 将把导出编排从 `app/ui/features/export_controller.py` 切换到 Preview Feature，旧路径保留兼容转发。
+- FGUI 图集、Spine/FFmpeg 适配和批量 Worker 先完成 ownership 收口，不重写外部工具算法，不改变 `output/fgui`、`output/video` 和角色图片输出。
+
+## Issue #44：Preview Feature 迁移完成（2026-08-24）
+
+- 导出编排已切换到 `app/features/preview/export_controller.py`；旧 `app/ui/features/export_controller.py` 改为兼容转发。
+- FGUI、Spine/FFmpeg 适配和批量/合成 Worker 已增加 Preview Feature 入口；外部工具参数、取消语义和 `output/character`、`output/fgui`、`output/video` 契约保持不变。
+- 已完成全量 pytest、Ruff、AST、实际 Qt 启动冒烟、文档同步和 diff 检查；下一阶段进入 Platform/Shared 收缩。
+
 ## 六、发布前规划（新功能使用 + 环境打包）
 
 > 详见 Issue #11，需双方讨论后再定实施顺序。
